third_party/blink/web_tests/http/tests/devtools/sources/debugger-breakpoints/reload-on-breakpoint-expected.txt

Tests that reloading while paused at a breakpoint doesn't execute code after the breakpoint.


Running: testFetchBreakpoint
Waiting for breakpoint.
Script execution paused.
Call stack:
    0) divergingFunction (reload-on-breakpoint.js:11)
    1)  (:1)
Reloading page...
Script execution resumed.
Page reloaded.
